refactor: model 폴더 생성, 관련 파일 이동 및 경로 수정
