+++ src/model/BridgeGame.js
@@ -1,6 +1,6 @@
-const BridgeMaker = require('./BridgeMaker');
-const BridgeRandomNumberGenerator = require('./BridgeRandomNumberGenerator');
-const { MOVE, MOVE_RESULT } = require('./constants/bridge');
+const BridgeMaker = require('../BridgeMaker');
+const BridgeRandomNumberGenerator = require('../BridgeRandomNumberGenerator');
+const { MOVE, MOVE_RESULT } = require('../constants/bridge');
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
